Update MockContract submit type int->float

+++ pdr_backend/predictoor/approach1/test/test_predictoor_agent1.py
@@ -89,7 +89,7 @@ def test_predictoor_agent1(monkeypatch):
             return S_PER_EPOCH
 
         def submit_prediction(
-            self, predval: bool, stake: int, timestamp: int, wait: bool = True
+            self, predval: bool, stake: float, timestamp: int, wait: bool = True
         ):  # pylint: disable=unused-argument
             if timestamp in self._prediction_slots:
                 print(f"      (Replace prev pred at time slot {timestamp})")
